+++ src/Calor.Compiler/Ast/AsyncNodes.cs
@@ -26,6 +26,5 @@ public sealed class AwaitExpressionNode : ExpressionNode
         ConfigureAwait = configureAwait;
     }
 
-    public override void Accept(IAstVisitor visitor) => visitor.Visit(this);
-    public override T Accept<T>(IAstVisitor<T> visitor) => visitor.Visit(this);
+
 }