+++ src/Calor.Compiler/Ast/CharOperationNode.cs
@@ -46,8 +46,7 @@ public sealed class CharOperationNode : ExpressionNode
         Arguments = arguments ?? throw new ArgumentNullException(nameof(arguments));
     }
 
-    public override void Accept(IAstVisitor visitor) => visitor.Visit(this);
-    public override T Accept<T>(IAstVisitor<T> visitor) => visitor.Visit(this);
+
 }
 
 /// <summary>