+++ src/Calor.Compiler/Ast/StringOperationNode.cs
@@ -158,8 +158,6 @@ public sealed class StringOperationNode : ExpressionNode
         ComparisonMode = comparisonMode;
     }
 
-    public override void Accept(IAstVisitor visitor) => visitor.Visit(this);
-    public override T Accept<T>(IAstVisitor<T> visitor) => visitor.Visit(this);
 
     /// <summary>
     /// Returns true if this operation supports string comparison modes.